route3py/pyproject.toml
[tool.poetry]
name = "route3py"
version = "0.1.0"
description = "api client for route3"
authors = ["Nico Mandery <nico@nmandery.net>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.38.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
grpcio-tools = "^1.38.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
